requirements.txt
openai
python-dotenv
flask
pydantic
